Warning fix

+++ Sources/FlitsGeohashC/flitsgeohash.c
@@ -123,7 +123,7 @@ GEOHASH_free_neighbors(GEOHASH_neighbors *neighbors)
 char*
 GEOHASH_get_adjacent(const char* hash, GEOHASH_direction dir)
 {
-    int len, idx;
+    size_t len, idx;
     const char *border_table, *neighbor_table;
     char *base, *refined_base, *ptr, last;
